Split Settings into editor Settings + tenant-admin Administration; rename /admin → /site-admin (#61)
* feat(security): implement feature gating and tenant plan info provider
refactor(services): replace ServiceTier usage with feature gate checks
fix(services): update quota enforcement to async methods

* feat: add quota management endpoints and frontend components

- Implemented new API endpoints for tenant quotas in QuotaEndpoints.cs.
- Updated ErrorResponses.cs to handle quota limit errors with long values.
- Added tests for quota-related functionality in various test files.
- Introduced new frontend components for displaying usage limits and storage monitoring.
- Refactored service tier handling to use lowercase values for consistency.
- Enhanced the API client to support fetching tenant quotas and managing subscription tiers.

* feat(quota): centralize quota and entitlement labels, and byte formatting helpers

* Refactor space management and testing

- Updated SpaceEndpoints to remove redundant parameters in UpdateAsync method.
- Added integration tests for ResourceGroupMemberRepository to ensure correct space-group membership behavior.
- Enhanced ResourceGroupMembersEditor component to handle 1:1 space-to-group assignments with confirmation before moving spaces.
- Removed GroupSpacesEditor component and its associated tests as part of the refactor.
- Updated ReportingApiPage tests to use consistent tier casing for membership.
- Improved SchedulerGrid component to use stable refs for edge-scroll functionality.
- Removed groupId from UpdateSpaceRequest interface to simplify space update logic.

* feat(settings): split Settings into editor-open Settings + tenant-admin Administration

Settings keeps the editor-open content tabs (Criteria, Templates, Presets,
Scheduling) and is now fully visible to editors and admins alike. Governance
tabs (Sites, Users, Organization, Configuration, Integrations, Usage & Limits)
move to a new /tenant-admin Administration page gated by a single
RequireTenantAdmin route guard, replacing the scattered per-tab isAdmin spreads.

Also renames the platform admin SPA route /admin -> /site-admin via a new
ROUTE_SITE_ADMIN constant so routes mirror their roles (site-admin <-> isSiteAdmin;
tenant-admin <-> tenant role==admin). The /api/admin/* backend namespace is
unchanged; only the SPA page route, navigation, and break-glass return target move.

Availability-event write affordances inside the editor-visible Scheduling tab stay
isAdmin-gated in the UI, since their backend writes are RequireAdminAccess.

Legacy ?tab= values and /settings/<moved> bookmarks redirect into /tenant-admin/*.


* test(auto-schedule): align tier mocks with lowercase ServiceTier type

useAutoScheduleAvailable compares tier against the lowercase ServiceTier
union ('free' | 'professional' | 'enterprise'), but these mocks supplied
capitalized 'Professional'/'Enterprise' that can never match — making three
assertions fail on main. Lowercase the mocks so they reflect the real
membership.tier contract. Test-only; the hook is correct.


* test(auth): update break-glass returnTo assertion to /site-admin

The AdminReturnPath was renamed from /admin to /site-admin in
AuthorizationExtensions.cs; the corresponding test assertion was missed.


---------

+++ backend/core/Helpers/TierLimitExceededException.cs
@@ -7,16 +7,16 @@ namespace Api.Helpers;
 public class QuotaExceededException : Exception
 {
     public string ResourceType { get; }
-    public int Limit { get; }
+    public long Limit { get; }
 
-    public QuotaExceededException(string resourceType, int limit)
+    public QuotaExceededException(string resourceType, long limit)
         : base($"Quota exceeded for {resourceType}. Maximum allowed: {limit}")
     {
         ResourceType = resourceType;
         Limit = limit;
     }
 
-    public QuotaExceededException(string resourceType, int limit, string message)
+    public QuotaExceededException(string resourceType, long limit, string message)
         : base(message)
     {
         ResourceType = resourceType;
@@ -31,7 +31,7 @@ public class TierLimitExceededException : QuotaExceededException
 {
     public string TierName { get; }
 
-    public TierLimitExceededException(string tierName, string resourceType, int limit)
+    public TierLimitExceededException(string tierName, string resourceType, long limit)
         : base(resourceType, limit, $"Service tier '{tierName}' limit exceeded for {resourceType}. Maximum allowed: {limit}")
     {
         TierName = tierName;
