+++ backend/src/Helpers/ErrorResponses.cs
@@ -83,7 +83,7 @@ public static class ErrorResponses
     /// Returns a 403 Forbidden response for a quota-limit hit. The frontend uses the
     /// resource type and limit to render "You've reached your X limit (limit/limit)."
     /// </summary>
-    public static IResult QuotaExceeded(string resourceType, int limit, string message)
+    public static IResult QuotaExceeded(string resourceType, long limit, string message)
     {
         return Results.Json(new ErrorResponse
         {
@@ -106,5 +106,5 @@ public record ErrorResponse
     /// <summary>Optional URL the frontend should navigate to after handling the error (e.g. "/admin" when a break-glass session ends).</summary>
     public string? ReturnTo { get; init; }
     /// <summary>Numeric limit associated with the error, e.g. the tier max for a <c>quota_exceeded</c> response.</summary>
-    public int? Limit { get; init; }
+    public long? Limit { get; init; }
 }
